Let code sniffer do a review
Github API changes forces us to place comments on pull request instead of commits which we did previously.
So converted code sniffer to place a review when it fails and place comments on Pull Request lines when they are inside the diff otherwise we just place it in main comment.